# docs/contracts/openapi/generated.json
{
  "components": {
    "schemas": {
...
      "CreateServiceAccount": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "environments": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Environments",
            "type": "array"
          },
          "expires_in_days": {
            "default": 365,
            "exclusiveMinimum": 0.0,
            "maximum": 3650.0,
            "title": "Expires In Days",
            "type": "integer"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Permissions",
            "type": "array"
          },
          "rate_limit_rps": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rate Limit Rps"
          },
          "slug": {
            "pattern": "^[a-z0-9][a-z0-9-]{1,62}$",
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "environments",
          "permissions"
        ],
        "title": "CreateServiceAccount",
        "type": "object"
      },
      "CreatedServiceAccount": {
        "additionalProperties": false,
        "description": "The one response that carries a key value (FR-PLAT-3).",
        "properties": {
          "account": {
            "$ref": "#/components/schemas/ServiceAccountView"
          },
          "key": {
            "description": "The full key. Shown once and never recoverable — nothing stores it.",
            "title": "Key",
            "type": "string"
          },
          "warning": {
            "default": "Store this now. It cannot be retrieved again.",
            "title": "Warning",
            "type": "string"
          }
        },
        "required": [
          "account",
          "key"
        ],
        "title": "CreatedServiceAccount",
        "type": "object"
      },
      "FieldError": {
        "additionalProperties": false,
        "description": "One field-level failure inside a problem response.\n\nPresent so that a rejected submission can be rendered against the form that produced\nit — FR-PLAT-11 requires deterministic failures to name the offending field rather\nthan only the operation.",
        "properties": {
          "code": {
            "description": "Stable machine code for this field.",
            "pattern": "^[A-Z][A-Z0-9_]*$",
            "title": "Code",
            "type": "string"
          },
          "field": {
            "description": "Dotted path to the offending field, e.g. 'spec.family'.",
            "title": "Field",
            "type": "string"
          },
          "message": {
            "description": "Human-readable explanation.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "code",
          "message"
        ],
        "title": "FieldError",
        "type": "object"
      },
...
      "KeyView": {
        "additionalProperties": false,
        "description": "A key as it can safely be shown: prefix and dates, never the secret.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "environment": {
            "title": "Environment",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "last_used_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "prefix": {
            "title": "Prefix",
            "type": "string"
          },
          "revoked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          }
        },
        "required": [
          "prefix",
          "environment",
          "created_at",
          "expires_at"
        ],
        "title": "KeyView",
        "type": "object"
      },
...
      "ProblemDetail": {
        "additionalProperties": false,
        "description": "An RFC 9457 problem response.\n\n`code` is the contract, not `title` or `status`: titles are prose and several\nconditions share a status, so a client that branches on anything else is branching on\nsomething the platform is free to reword.",
        "properties": {
          "code": {
            "description": "Stable machine-readable code, namespaced per module and enumerated in that module's spec. This is what clients branch on.",
            "pattern": "^[A-Z][A-Z0-9_]*$",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Explanation specific to this occurrence.",
            "title": "Detail"
          },
          "errors": {
            "default": [],
            "description": "Field-level errors, where the failure is field-attributable.",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Errors",
            "type": "array"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Path of the failing request.",
            "title": "Instance"
          },
          "status": {
            "description": "HTTP status code.",
            "maximum": 599.0,
            "minimum": 100.0,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Short human-readable summary, stable per type.",
            "title": "Title",
            "type": "string"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "OpenTelemetry trace id for this request. Appears in every log line for the request, so a support conversation starts with an identifier (FR-PLAT-42).",
            "title": "Trace Id"
          },
          "type": {
            "default": "about:blank",
            "description": "URI identifying the problem type; documentation lives at that URI.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "code"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
...
      "ServiceAccountView": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "environments": {
            "items": {
              "type": "string"
            },
            "title": "Environments",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "keys": {
            "items": {
              "$ref": "#/components/schemas/KeyView"
            },
            "title": "Keys",
            "type": "array"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          },
          "rate_limit_rps": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rate Limit Rps"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "description",
          "environments",
          "permissions",
          "rate_limit_rps",
          "keys"
        ],
        "title": "ServiceAccountView",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "GI Pricing Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/service-accounts": {
      "post": {
        "description": "Create the account and its first key. The key is in this response and nowhere else.",
        "operationId": "create_service_account_api_v1_service_accounts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceAccount"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedServiceAccount"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not permitted"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Create a service account",
        "tags": [
          "platform"
        ]
      }
    },
    "/api/v1/service-accounts/{account_id}/keys/{prefix}": {
      "delete": {
        "operationId": "revoke_key_api_v1_service_accounts__account_id__keys__prefix__delete",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "prefix",
            "required": true,
            "schema": {
              "title": "Prefix",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceAccountView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not permitted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Revoke a key immediately",
        "tags": [
          "platform"
        ]
      }
    },
    "/api/v1/service-accounts/{account_id}/rotate": {
      "post": {
        "description": "Issue a new key and give the old one a deadline (FR-PLAT-3).\n\nThe old key keeps working for the overlap window rather than dying immediately. A\nrotation that breaks production the instant it is requested is a rotation nobody\nperforms, and an unrotated key is the failure this exists to prevent.",
        "operationId": "rotate_key_api_v1_service_accounts__account_id__rotate_post",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "overlap_days",
            "required": false,
            "schema": {
              "default": 7,
              "maximum": 90,
              "minimum": 0,
              "title": "Overlap Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedServiceAccount"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not permitted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Rotate the key with an overlap window",
        "tags": [
          "platform"
        ]
      }
    },
...
  }
}